+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
 -   Fix issue where query parameters with multiple of the same key were not processing more than one of it's values
+-   Notify and return an error if the response body provided to RespondHTTP is too large